pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf-accessibility-platform"
version = "0.1.0"
description = "Bootstrap service for a PDF accessibility remediation platform."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
  "fastapi>=0.115,<1",
  "orjson>=3.10,<4",
  "pdfplumber>=0.11,<1",
  "pikepdf>=10,<11",
  "pydantic-settings>=2.6,<3",
  "PyMuPDF>=1.27,<2",
  "pypdf>=5,<7",
  "pytesseract>=0.3.13,<1",
  "python-multipart>=0.0.18,<1",
  "prometheus-client>=0.21,<1",
  "structlog>=25,<26",
  "uvicorn>=0.30,<1",
  "boto3>=1.34,<2",
  "openai>=1.0,<2",
  "anthropic>=0.30,<1",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27,<1",
  "pytest>=8,<9",
  "pytest-asyncio>=1,<2",
]

[project.scripts]
pdf-accessibility-api = "pdf_accessibility.main:run"
pdf-accessibility-cli = "pdf_accessibility.cli.main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
